Move HTTP Transport from examples/ into transport.py
